% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pakret-package.R
\docType{package}
\name{pakret-package}
\alias{pakret}
\alias{pakret-package}
\title{pakret: Cite R Packages on the Fly}
\description{
References and cites 'R' packages on the fly in 'R Markdown' and
'Quarto'. 'pakret' provides a minimalist 'API' that generates preformatted
citations of 'R' packages and other 'R' objects, and adds their reference
to a '.bib' file directly from within your document.
}
\author{
\strong{Maintainer}: Arnaud Gallou \email{arangacas@gmail.com} (\href{https://orcid.org/0000-0002-1002-4247}{ORCID}) [copyright holder]

}
\keyword{internal}
