% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkrt.R
\name{pkrt}
\alias{pkrt}
\title{Cite R or an R package}
\usage{
pkrt(pkg)
}
\arguments{
\item{pkg}{Name of the package to cite. You can use \code{"R"} as an alias for the
base package to cite R.}
}
\value{
A character string with S3 class \code{AsIs}.
}
\description{
Creates a preformatted citation for R or an R package. This
function is normally used within an R Markdown or Quarto document, where
\code{pkrt()} automatically references the cited package to the first (by
default) \code{.bib} file specified in the YAML header if no reference for the
package already exists.
}
\examples{
pkrt("pakret")

pkrt("R")
}
