% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pakret-package.R
\docType{package}
\name{pakret-package}
\alias{pakret}
\alias{pakret-package}
\title{pakret: Cite R Packages on the Fly in R Markdown and Quarto}
\description{
References and cites 'R' and 'R' packages on the fly in 'R Markdown' and 'Quarto'. 'pakret' provides a minimalistic 'API' that generates preformatted citations of 'R' and 'R' packages, and adds their reference to a '.bib' file directly from within your document.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arnaudgallou/pakret}
  \item \url{https://arnaudgallou.github.io/pakret/}
  \item Report bugs at \url{https://github.com/arnaudgallou/pakret/issues}
}

}
\author{
\strong{Maintainer}: Arnaud Gallou \email{arangacas@gmail.com} (\href{https://orcid.org/0000-0002-1002-4247}{ORCID}) [copyright holder]

}
\keyword{internal}
