% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkrt.R
\name{pkrt}
\alias{pkrt}
\title{Cite R or an R package}
\usage{
pkrt(pkg)
}
\arguments{
\item{pkg}{Name of the package to cite.}
}
\value{
A character string.
}
\description{
Creates a preformatted citation of R or an R package. This
function should normally only be used in an R Markdown or Quarto document,
in which case \code{pkrt()} automatically references the cited package in the
first (by default) \code{.bib} file specified in the YAML header if no
references of the package already exist.
}
\examples{
pkrt("pakret")

pkrt("R")
}
